Remove fractional resource consumption (#4061)
Replaced probabilistic resource consumption with deterministic
consumption by changing `InventoryProbability` (float) to
`InventoryQuantity` (int) throughout the codebase.

This is part of removing floats from the simulation. For features that
are in active use, I'm aiming to replace floating calculations with
integer calculations. For features that aren't in active use, I'm just
removing the feature.


https://app.asana.com/1/1209016784099267/project/1209096222434381/task/1211547145607801?focus=true

---------

+++ packages/mettagrid/python/src/mettagrid/config/mettagrid_config.py
@@ -97,7 +97,7 @@ class ActionConfig(Config):
     enabled: bool = Field(default=True)
     # required_resources defaults to consumed_resources. Otherwise, should be a superset of consumed_resources.
     required_resources: dict[str, int] = Field(default_factory=dict)
-    consumed_resources: dict[str, float] = Field(default_factory=dict)
+    consumed_resources: dict[str, int] = Field(default_factory=dict)
 
     def actions(self) -> list[Action]:
         if self.enabled:
